# perf(sqlalchemy): lazily construct the inspector object