--- a/pyproject.toml
+++ b/pyproject.toml
@@ -88,7 +88,7 @@ snowflake-connector-python = { version = ">=2.7.10,<4", optional = true }
 snowflake-sqlalchemy = { version = ">=1.4.1,<2", optional = true }
 sqlalchemy = { version = ">=1.4,<3", optional = true }
 sqlalchemy-views = { version = ">=0.3.1,<1", optional = true }
-trino = { version = ">=0.319,<1", optional = true, extras = ["sqlalchemy"] }
+trino = { version = ">=0.321,<1", optional = true, extras = ["sqlalchemy"] }
 
 [tool.poetry.group.dev.dependencies]
 black = ">=22.1.0,<24"