This PR speeds up connecting to backends where `sqlalchemy.inspect` executes a
significant amount of code/traverses the network.

Snowflake is one such backend.

Before:

```
In [1]: from ibis.interactive import *

In [2]: import os

In [3]: %time con = ibis.connect(os.environ["SNOWFLAKE_URL"])
CPU times: user 574 ms, sys: 56.6 ms, total: 630 ms
Wall time: 1.88 s
```

After:

```
In [1]: from ibis.interactive import *

In [2]: import os

In [3]: %time con = ibis.connect(os.environ["SNOWFLAKE_URL"])
CPU times: user 475 ms, sys: 36.9 ms, total: 512 ms
Wall time: 514 ms
```